requirements.txt
Django==3.2.5
djangorestframework==3.12.4
gunicorn==20.1.0
dj-database-url==0.5.0
psycopg2==2.9.1
